+++ src/finalisers/cjs.ts
@@ -51,7 +51,13 @@ export default function cjs(
 		intro += importBlock + '\n\n';
 	}
 
-	const exportBlock = getExportBlock(exports, dependencies, exportMode, 'module.exports =');
+	const exportBlock = getExportBlock(
+		exports,
+		dependencies,
+		exportMode,
+		options.interop,
+		'module.exports ='
+	);
 
 	magicString.prepend(intro);