+++ src/finalisers/umd.ts
@@ -113,7 +113,7 @@ export default function umd(
 
 	if (intro) magicString.prepend(intro);
 
-	const exportBlock = getExportBlock(exports, dependencies, exportMode);
+	const exportBlock = getExportBlock(exports, dependencies, exportMode, options.interop);
 	if (exportBlock) magicString.append('\n\n' + exportBlock);
 	if (exportMode === 'named' && options.legacy !== true)
 		magicString.append(`\n\n${esModuleExport}`);