+++ test/form/samples/interop-false-reexport/_config.js
@@ -1,0 +1,6 @@
+module.exports = {
+	description: 'reexports with interop: false',
+	options: {
+		output: { name: 'foo', interop: false },
+	}
+};