+++ test/form/samples/interop-false-reexport/_expected/cjs.js
@@ -1,0 +1,10 @@
+'use strict';
+
+Object.defineProperty(exports, '__esModule', { value: true });
+
+var external = require('external');
+
+
+
+exports.q = external.p;
+exports.p = external.default;