+++ test/form/samples/interop-false-reexport/_expected/iife.js
@@ -1,0 +1,11 @@
+var foo = (function (exports,external) {
+	'use strict';
+
+
+
+	exports.q = external.p;
+	exports.p = external.default;
+
+	return exports;
+
+}({},external));